+++ src/Entities/Schema.php
@@ -39,12 +39,6 @@ final class Schema implements IteratorAggregate, ArrayAccess
             'bank-accounts/{bankAccountId}' => function (array $content) {
                 return new BankAccount($content);
             },
-            'blacklists' => function (array $content) {
-                return new Collection(new Blacklist(), $content);
-            },
-            'blacklists/{blacklistId}' => function (array $content) {
-                return new Blacklist($content);
-            },
             'blocklists' => function (array $content) {
                 return new Collection(new Blocklist(), $content);
             },