+++ src/Client.php
@@ -43,7 +43,6 @@ use RuntimeException;
  * @method Services\AuthenticationOptionsService authenticationOptions()
  * @method Services\AuthenticationTokenService authenticationTokens()
  * @method Services\BankAccountService bankAccounts()
- * @method Services\BlacklistService blacklists()
  * @method Services\BlocklistService blocklists()
  * @method Services\CustomerCredentialService customerCredentials()
  * @method Services\CustomerService customers()
@@ -109,7 +108,6 @@ final class Client
         'authenticationOptions' => Services\AuthenticationOptionsService::class,
         'authenticationTokens' => Services\AuthenticationTokenService::class,
         'bankAccounts' => Services\BankAccountService::class,
-        'blacklists' => Services\BlacklistService::class,
         'blocklists' => Services\BlocklistService::class,
         'customerCredentials' => Services\CustomerCredentialService::class,
         'customers' => Services\CustomerService::class,