+++ tests/Api/ServiceTest.php
@@ -1259,6 +1259,11 @@ class ServiceTest extends BaseTestCase
                 Entities\ExperianCredential::class,
             ],
             [
+                'taxjarCredentials',
+                Services\TaxJarCredentialsService::class,
+                Entities\TaxJarCredential::class,
+            ],
+            [
                 'webhooksTracking',
                 Services\WebhookTrackingService::class,
                 Entities\WebhookTracking::class,